Delete 'impl_device.hpp' and move `*_implementation` into respective folders
