+++ device/CMakeLists.txt
@@ -1,18 +1,18 @@
 
 set(UMD_DEVICE_SRCS
     architecture_implementation.cpp
-    blackhole_implementation.cpp
     cpuset_lib.cpp
-    grayskull_implementation.cpp
     tlb.cpp
     tt_cluster_descriptor.cpp
     tt_device.cpp
     tt_silicon_driver.cpp
     tt_silicon_driver_common.cpp
     tt_soc_descriptor.cpp
-    wormhole_implementation.cpp
     simulation/tt_simulation_device.cpp
     simulation/tt_simulation_host.cpp
+    blackhole/blackhole_implementation.cpp
+    grayskull/grayskull_implementation.cpp
+    wormhole/wormhole_implementation.cpp
 )
 add_library(umd_device SHARED ${UMD_DEVICE_SRCS})
 target_link_libraries(umd_device
