+++ device/blackhole/blackhole_implementation.cpp
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#include "device/blackhole_implementation.h"
+#include "blackhole_implementation.h"
 
 namespace tt::umd {
