+++ device/blackhole/blackhole_implementation.h
