+++ device/wormhole/wormhole_implementation.cpp
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#include "device/wormhole_implementation.h"
+#include "wormhole_implementation.h"
 
 namespace tt::umd {
